add tests, fix #1, move kvstore into files
