+++ CMakeLists.txt
@@ -17,9 +17,9 @@ include(cmake/Git.cmake)
 ### SETTINGS ###
 
 # add all headers (.h, .hpp) to this
-set(PRJ_HEADERS )
+set(PRJ_HEADERS src/KVStore.h)
 # add all source files (.cpp) to this, except the one with main()
-set(PRJ_SOURCES )
+set(PRJ_SOURCES src/KVStore.cpp)
 # set the source file containing main()
 set(PRJ_MAIN src/main.cpp)
 # set the source file containing the test's main
